+++ examples/harbor-crud-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # harbor-crud-api
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @velajs/crud@1.18.1
+  - @velajs/crud-memory@1.18.1
+
 ## 0.0.1
 
 ### Patch Changes